codechef; use `int` instead of `short`
Most likely the `int` is faster and the saved memory is not worth.

+++ codechef/easy/life_the_universe_and_everything.cpp
@@ -12,7 +12,7 @@ int main()
 {
     use_io_optimizations();
 
-    short input_number;
+    int input_number;
 
     while (cin >> input_number && input_number != 42)
     {
